requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
tensorflow==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.20
pillow==11.1.0